skip doing EC math for Armory wallets
Wallets created by recent version of Armory (v0.90+, c. Nov 2013)
deterministically derive their chaincode from their master private key
(and the chaincode is not encrypted). This can be used to bypass doing
elliptic curve math, however in practice this offers only a very tiny
speedup. Given this, plus that it's not possible to determine which
Armory version created the wallet, this branch likely won't be merged.